refactor rtk 오류 수정
