+++ src/App.jsx
@@ -1,7 +1,7 @@
 // import ProductDetail from "./pages/ProductDetail";
 import {Routes, Route} from "react-router-dom";
 import Layout from "./components/templates/Layout";
-import MyPage from "./pages/myPage/MyPage";
+import MyPage from "./components/pages/myPage/MyPage";
 import Main from "./components/pages/Main";
 import CategoryDetail from "./components/pages/CategoryDetail";
 import ProductDetail from "./components/pages/ProductDetail";
