--- src/components/organisms/Form.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Form = () => {
-    return <div>Form</div>;
-};
-
-export default Form;
