+++ src/components/organisms/ProductActions.jsx
@@ -1,5 +1,10 @@
 import Button from "../atoms/Button";
-const ProductActions = ({addfavoriteProducts}) => {
+const ProductActions = () => {
+
+    const addfavoriteProducts = () => {
+        console.log(1)
+    }
+    
     return (
         <div className="grid grid-cols-3 grid-rows-1 gap-4">
             <Button
